requirements.txt
certifi==2016.9.26
Django==1.10.5
django-filter==1.0.1
django-friendship==1.5.0
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
drfdocs==0.0.11
Markdown==2.6.7
olefile==0.44
Pillow==4.0.0
psycopg2==2.7.3
PyJWT==1.4.2
numpy==1.13.3
scipy==1.0.0
nltk==3.2.5
gensim==3.1.0